+++ src/models/tbl_store.js
@@ -17,7 +17,7 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false
     },
     storeParentId: {
-      type: DataTypes.INTEGER(10),
+      type: DataTypes.INTEGER,
       allowNull: true
     },
     address01: {
@@ -29,7 +29,7 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: true
     },
     // cityId: {
-    //   type: DataTypes.INTEGER(10),
+    //   type: DataTypes.INTEGER,
     //   allowNull: false,
     // },
     state: {
