+++ src/models/tbl_transfer_out.js
@@ -8,11 +8,11 @@ module.exports = function (sequelize, DataTypes) {
             primaryKey: true
         },
         storeId: {
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.INTEGER,
             allowNull: false,
         },
         storeIdReceiver: {
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.INTEGER,
             allowNull: false,
         },
         transNo: {
@@ -32,7 +32,7 @@ module.exports = function (sequelize, DataTypes) {
             allowNull: false
         },
         employeeId: {
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.INTEGER,
             allowNull: false
         },
         carNumber: {
@@ -57,7 +57,7 @@ module.exports = function (sequelize, DataTypes) {
             allowNull: true
         },
         totalPackage: {
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.INTEGER,
             allowNull: true
         },
         description: {
