+++ src/models/view/vw_transfer_in.js
@@ -9,13 +9,13 @@ module.exports = function (sequelize, DataTypes) {
       primaryKey: true
     },
     storeId: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
     storeName: { type: DataTypes.STRING, allowNull: true },
     storeCode: { type: DataTypes.STRING, allowNull: true },
     storeIdSender: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
     storeNameSender: { type: DataTypes.STRING, allowNull: true },
@@ -40,7 +40,7 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false
     },
     employeeId: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER,
       allowNull: false
     },
     employeeName: { type: DataTypes.STRING, allowNull: true },
@@ -64,7 +64,7 @@ module.exports = function (sequelize, DataTypes) {
       }
     },
     totalPackage: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER,
       allowNull: true
     },
     description: {
